Refactor pydantic move core types to ucl_open.core
