+++ src/ucl_open/core/base.py
@@ -14,7 +14,6 @@ class BaseSchema(BaseModel):
         from_attributes=True,
     )
 
-
 class Device(BaseSchema):
     """The base class for creating Aeon hardware device models."""
