+++ src/ucl_open/core/task.py
@@ -1,7 +1,8 @@
 from typing import Optional
 from pydantic import Field
 
-from ucl_open.base import BaseSchema
+from ucl_open.core.base import BaseSchema
+
 
 class TaskParameters(BaseSchema):
     rng_seed: Optional[float] = Field(default=None, description="Seed of the random number generator for these task parameters")
